Added install and uninstall listeners, and the COIoshelper*.dlls
The intention was to have this, together with a dependency to the
izpack-native.dll, make the installed service appear in the "Programs
and features" setting of windows.

It didn't work, possibly because the settings added to the
izpack.xml.ftl file, are izpack 5 setting and the
windows-service-installer maven plugin uses izpack 4.